Don't require Unpin when pin projecting.
There are two scenarios here. Either you want users to be able to pass in
streams that are `!Unpin`, in which case pin_project and the compiler
have your back and make sure this works, or...

You are going to require Unpin on it in which case you can remove the pin
projection.

Both together don't make sense. Either one works here, but by removing the
Unpin the API is less restrictive, which I would favor, even if I don't
immediately see what concrete `!Unpin` type someone might want to call this with,
but sooner or later someone probably will, so why disable if it can work.

This leaves another problem though, there are many other places in the lib
where `Unpin` is required, notably on the wrappers around `Stream`, so it's
not said that with this change users can actually use any `!Unpin` streams.

There is one other use of pin-project on `MidHandshake`. As far as I can
tell the effect of this one is that `Role::InternalStream` can now be `!Unpin`
in this specific file, but is it even possible to use the lib with an `!Unpin`
type for this?

It looks like it might be good to rethink the choices about pin for the
entire lib and make something consistent. Maybe aim for the largest API
compatibility by trying to never require `Unpin` and seeing if pin-project
makes that possible.

In any case, the Unpin in `Stream` has no good reason for being there unless
you would like to [get rid of pin-project](https://github.com/najamelan/tokio-tungstenite/commit/18cc347aaa7f106b28f54ed38b0a030130ffbc1f).

As an added bonus, if you pin project, there is no need for re-pinning in the match.